WL-0MLARH59Q0FY64WN: extract TUI controller
